requirements.txt
numpy
pandas
matplotlib
scikit-learn
seaborn
scipy
flask
pillow
jupyter
